+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 * Added support for providing certificate options with HTTPS requests. See further documentation: [https](https://nodejs.org/api/https.html#https_https_request_options_callback) and [tls](https://nodejs.org/api/tls.html#tls_tls_connect_options_callback).
 * `bodyOnly` options has been removed. This means by all responses will be of type `Response` (`statusCode`, `headers` and `body`).
-* Added **TypeScript** declaration.
 * Changed function name from `configureGlobalAgent()` to `globalAgent()`.
+* Can handle file uploads. Pass a read `stream` to `options.body`, or set a file path in `options.path`.
+* Added **TypeScript** declaration.
 
 ## 0.4.0